Split up some files to decrease readability
